# docs/swagger.yaml
basePath: /
definitions:
  handler.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  handler.GenerateAPIKeyRequest:
    properties:
      expires_in_days:
        type: integer
      name:
        type: string
    type: object
...
  handler.LoginWithMagicLinkRequest:
    properties:
      token:
        type: string
    type: object
...
  handler.RequestMagicLinkRequest:
    properties:
      email:
        type: string
    type: object
  handler.UpdateProfileRequest:
    properties:
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A production-ready authentication microservice with modern features
    including 2FA, magic links, session management, and API keys.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Authentication Service API
  version: "1.0"
paths:
...
  /api-keys:
    get:
      consumes:
      - application/json
      description: Returns all API keys for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get API keys
      tags:
      - api-keys
  /api-keys/generate:
    post:
      consumes:
      - application/json
      description: Generates a new API key for service-to-service authentication
      parameters:
      - description: API key generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GenerateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate API key
      tags:
      - api-keys
  /auth/magic-link:
    post:
      consumes:
      - application/json
      description: Authenticates using a magic link token from email
      parameters:
      - description: Magic link login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginWithMagicLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login with magic link
      tags:
      - magic-link
  /auth/magic-link/request:
    post:
      consumes:
      - application/json
      description: Sends a passwordless login link to the user's email
      parameters:
      - description: Magic link request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RequestMagicLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Request magic link
      tags:
      - magic-link
...
  /login-history:
    get:
      consumes:
      - application/json
      description: Returns login history for the authenticated user
      parameters:
      - description: 'Number of records to return (default: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get login history
      tags:
      - account
...
  /password/change:
    post:
      consumes:
      - application/json
      description: Changes the user's password (requires current password)
      parameters:
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - account
...
  /profile:
    put:
      consumes:
      - application/json
      description: Updates the user's profile information
      parameters:
      - description: Profile update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - account
...
  /sessions:
    get:
      consumes:
      - application/json
      description: Returns all active sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active sessions
      tags:
      - sessions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"